refact(create): moved code in modules
subj
